feat: 添加雪花算法生成器 feat: 添加SnowFlowIntUtil类实现雪花算法 feat: 添加MockDataUtil工具类用于生成模拟图片数据 feat: 添加MockDataImageReq请求实体类 feat: 添加MockDataController控制器用于处理模拟图片请求 feat: 更新BusinessConfig配置类 feat: 更新Constant常量类

+++ spring-business/src/main/java/org/start2do/dto/MenuRespV2.java
@@ -16,7 +16,7 @@ public class MenuRespV2 {
 
     private Object value2;
 
-    private Short sort;
+    private Integer sort;
 
     public MenuRespV2(String label, Object value, Object value2) {
         this.label = label;
